+++ api/src/raw/engine.py
@@ -3,13 +3,17 @@ import os
 from pathlib import Path
 
 from llama_index.core import (
+    Settings,
     SimpleDirectoryReader,
     StorageContext,
     VectorStoreIndex,
     load_index_from_storage,
 )
+from llama_index.core.node_parser import SimpleNodeParser
 from llama_index.vector_stores.qdrant import QdrantVectorStore
 from qdrant_client import AsyncQdrantClient, QdrantClient, models
+
+from raw.ollama import Ollama
 
 
 def get_index():
@@ -31,6 +35,29 @@ def get_index():
         )
 
     return index
+
+
+def init_settings():
+    Settings.llm = Ollama(
+        model="mixtral:latest",
+        base_url="https://mirage.kite.ume.de/ollama",
+        request_timeout=240,
+        temperature=0,
+    )
+
+    node_parser = SimpleNodeParser.from_defaults(chunk_size=512, chunk_overlap=32)
+    Settings.embed_model = "local:BAAI/bge-small-en-v1.5"
+    Settings.node_parser = node_parser
+
+
+def get_llm():
+    return Settings.llm
+
+
+def get_chat_engine():
+    return get_index().as_chat_engine(
+        similarity_top_k=3, chat_mode="condense_plus_context", use_async=True
+    )
 
 
 def load_documents(data_path):
@@ -69,6 +96,7 @@ def update_index(index, data_path):
     updated = index.refresh_ref_docs(docs)
     for doc, is_new in zip(docs, updated):
         print(doc.get_doc_id(), f"Updated: {is_new}")
+    index.storage_context.persist(persist_dir="./storage")
 
 
 def delete_index(index):
